.specify/templates/platform/tests/requirements.txt
copier>=9.4.0,<10
pytest>=8.0
pyyaml>=6.0
